Move sidebar anchor tests to Swift Testing file
